many many code improvements

+++ elements/vector.cpp
@@ -3,8 +3,8 @@
 #include <iostream>
 #include "vector.h"
 
-using std::tanf;
-using std::atanf;
+using std::tan;
+using std::atan;
 using std::abs;
 using std::round;
 
@@ -25,7 +25,7 @@ Vector::Vector(Dot _begin, float _angle, int endY)
 float Vector::CountAngle() {
     int opposCatet = end.x - begin.x;
     int contCatet = begin.y - end.y;
-    angle = atanf(static_cast<float>(opposCatet)/static_cast<float>(contCatet));
+    angle = atan(static_cast<float>(opposCatet)/static_cast<float>(contCatet));
     angle *= 180/static_cast<float>(M_PI);
     return angle;
 }
